+++ vector/src/main/java/im/vector/riotx/features/login/LoginResetPasswordFragment.kt
@@ -25,10 +25,10 @@ import com.airbnb.mvrx.Loading
 import com.airbnb.mvrx.Success
 import com.jakewharton.rxbinding3.widget.textChanges
 import im.vector.riotx.R
-import im.vector.riotx.core.error.ErrorFormatter
 import im.vector.riotx.core.extensions.hideKeyboard
 import im.vector.riotx.core.extensions.isEmail
 import im.vector.riotx.core.extensions.showPassword
+import im.vector.riotx.core.extensions.toReducedUrl
 import io.reactivex.Observable
 import io.reactivex.functions.BiFunction
 import io.reactivex.rxkotlin.subscribeBy
@@ -38,9 +38,7 @@ import javax.inject.Inject
 /**
  * In this screen, the user is asked for email and new password to reset his password
  */
-class LoginResetPasswordFragment @Inject constructor(
-        private val errorFormatter: ErrorFormatter
-) : AbstractLoginFragment() {
+class LoginResetPasswordFragment @Inject constructor() : AbstractLoginFragment() {
 
     private var passwordShown = false
 
@@ -57,7 +55,7 @@ class LoginResetPasswordFragment @Inject constructor(
     }
 
     private fun setupUi(state: LoginViewState) {
-        resetPasswordTitle.text = getString(R.string.login_reset_password_on, state.homeServerUrlSimple)
+        resetPasswordTitle.text = getString(R.string.login_reset_password_on, state.homeServerUrl.toReducedUrl())
     }
 
     private fun setupSubmitButton() {
@@ -136,14 +134,6 @@ class LoginResetPasswordFragment @Inject constructor(
 
     override fun resetViewModel() {
         loginViewModel.handle(LoginAction.ResetResetPassword)
-    }
-
-    override fun onError(throwable: Throwable) {
-        AlertDialog.Builder(requireActivity())
-                .setTitle(R.string.dialog_title_error)
-                .setMessage(errorFormatter.toHumanReadable(throwable))
-                .setPositiveButton(R.string.ok, null)
-                .show()
     }
 
     override fun updateWithState(state: LoginViewState) {