+++ vector/src/main/java/im/vector/riotx/features/login/terms/LoginTermsFragment.kt
@@ -19,13 +19,12 @@ package im.vector.riotx.features.login.terms
 import android.os.Bundle
 import android.os.Parcelable
 import android.view.View
-import androidx.appcompat.app.AlertDialog
 import butterknife.OnClick
 import com.airbnb.mvrx.args
 import im.vector.riotx.R
-import im.vector.riotx.core.error.ErrorFormatter
 import im.vector.riotx.core.extensions.cleanup
 import im.vector.riotx.core.extensions.configureWith
+import im.vector.riotx.core.extensions.toReducedUrl
 import im.vector.riotx.core.utils.openUrlInExternalBrowser
 import im.vector.riotx.features.login.AbstractLoginFragment
 import im.vector.riotx.features.login.LoginAction
@@ -44,8 +43,7 @@ data class LoginTermsFragmentArgument(
  * LoginTermsFragment displays the list of policies the user has to accept
  */
 class LoginTermsFragment @Inject constructor(
-        private val policyController: PolicyController,
-        private val errorFormatter: ErrorFormatter
+        private val policyController: PolicyController
 ) : AbstractLoginFragment(),
         PolicyController.PolicyControllerListener {
 
@@ -106,16 +104,8 @@ class LoginTermsFragment @Inject constructor(
         loginViewModel.handle(LoginAction.AcceptTerms)
     }
 
-    override fun onError(throwable: Throwable) {
-        AlertDialog.Builder(requireActivity())
-                .setTitle(R.string.dialog_title_error)
-                .setMessage(errorFormatter.toHumanReadable(throwable))
-                .setPositiveButton(R.string.ok, null)
-                .show()
-    }
-
     override fun updateWithState(state: LoginViewState) {
-        policyController.homeServer = state.homeServerUrlSimple
+        policyController.homeServer = state.homeServerUrl.toReducedUrl()
         renderState()
     }